+++ app/blog/page.tsx
@@ -1,0 +1,27 @@
+import ListLayout from '@/layouts/MDX/ListLayout';
+import MainLayout from '@/layouts/MainLayout';
+import { sortedBlogPost } from '@/lib/utils/contentlayer';
+import { allBlogs } from 'contentlayer/generated';
+
+export const POSTS_PER_PAGE = 5;
+
+export default function Blog() {
+  const activePosts = allBlogs.filter((p) => p.draft === false);
+  const posts = sortedBlogPost(activePosts);
+  const initialDisplayPosts = posts.slice(0, POSTS_PER_PAGE);
+  const pagination = {
+    currentPage: 1,
+    totalPages: Math.ceil(posts.length / POSTS_PER_PAGE),
+  };
+
+  return (
+    <MainLayout>
+      <ListLayout
+        posts={posts}
+        initialDisplayPosts={initialDisplayPosts}
+        pagination={pagination}
+        title="Blog"
+      />
+    </MainLayout>
+  );
+}
