+++ lib/spotify.ts
@@ -30,6 +30,9 @@ export const getNowPlaying = async () => {
     headers: {
       Authorization: `Bearer ${access_token}`,
     },
+    next: {
+      revalidate: 30,
+    },
   });
 };
