+++ pages/blog/page/[page].tsx
@@ -1,12 +1,12 @@
 import { PageSEO } from '@/components/SEO';
 import siteMetadata from '@/data/siteMetadata';
-import MainLayout from '@/layouts/MainLayout';
 import ListLayout from '@/layouts/MDX/ListLayout';
+import MainLayout from '@/layouts/MainLayout';
 import { allCoreContent } from '@/lib/utils/contentlayer';
 import { allBlogs } from 'contentlayer/generated';
 import { InferGetStaticPropsType } from 'next';
 import { sortedBlogPost } from '../../../lib/utils/contentlayer';
-import { POSTS_PER_PAGE } from '../../blog';
+import { POSTS_PER_PAGE } from '../../_blog';
 
 export const getStaticPaths = async () => {
   const totalPosts = sortedBlogPost(allBlogs);
